Cover the Evolve provider ordering with behavioral tests
Extract the shared provider registry (metadata, ordering, marks) into its own light module so the chat picker, Manage models, and SettingsView read one source of truth without importing the heavy panel, and replace the new implementation-text assertions with behavioral coverage that renders the exported registry.
