+++ frontend/src/components/ChatView/__tests__/modelProviderOrder.test.js
@@ -1,33 +1,32 @@
 import test from 'node:test'
 import assert from 'node:assert/strict'
-import { readFileSync } from 'node:fs'
+import { createElement } from 'react'
+import { renderToStaticMarkup } from 'react-dom/server'
 
-const panel = readFileSync(new URL('../ChatSettingsPanel.jsx', import.meta.url), 'utf8')
-const panelCss = readFileSync(new URL('../ChatSettingsPanel.css', import.meta.url), 'utf8')
-const manageModels = readFileSync(new URL('../ManageModelsModal.jsx', import.meta.url), 'utf8')
+import { PROVIDER_ORDER, PROVIDER_INFO } from '../providerRegistry.jsx'
 
-test('chat model surfaces list Möbius after Codex and Claude', () => {
-  assert.match(
-    panel,
-    /export const PROVIDER_ORDER = \['codex', 'claude', 'mobius'\]/,
-  )
-  assert.match(panel, /mobius:\s*\{[\s\S]*label: 'Möbius subscription'/)
-  assert.doesNotMatch(panel, /label: 'Möbius trial'/)
+test('the app-owned Möbius subscription follows the connected coding subscriptions', () => {
+  assert.deepEqual(PROVIDER_ORDER, ['codex', 'claude', 'mobius'])
 })
 
-test('Möbius models reuse the Android notification mark in the shell ink color', () => {
-  assert.match(panel, /className="csp__mobius-logo"/)
-  assert.doesNotMatch(panel, /M8\.2 7\.2c-2\.9/)
-  assert.match(
-    panelCss,
-    /mask:\s*url\(['"]\/icons\/notification-badge\.svg['"]\)[^;]*;/,
-  )
-  assert.match(panelCss, /\.csp__mobius-logo\s*{[\s\S]*background:\s*currentColor;/)
+test('every ordered provider carries product-facing metadata and Möbius reads as a subscription', () => {
+  for (const id of PROVIDER_ORDER) {
+    const info = PROVIDER_INFO[id]
+    assert.ok(info, `${id} has provider metadata`)
+    assert.equal(typeof info.label, 'string')
+    assert.equal(typeof info.Logo, 'function')
+  }
+  assert.equal(PROVIDER_INFO.mobius.label, 'Möbius subscription')
+  // The trial is activated silently on sign-in; the picker never brands
+  // the app-owned provider as a "trial".
+  for (const info of Object.values(PROVIDER_INFO)) {
+    assert.notEqual(info.label, 'Möbius trial')
+  }
 })
 
-test('the public Evolve row does not expose its internal wire id', () => {
-  assert.match(
-    manageModels,
-    /pid !== 'mobius' && \([\s\S]*className="mmm-row__sub"/,
-  )
+test('the Möbius mark renders as the shared ink glyph, not an inlined provider path', () => {
+  const markup = renderToStaticMarkup(createElement(PROVIDER_INFO.mobius.Logo))
+  assert.match(markup, /class="csp__mobius-logo"/)
+  // No inlined OpenAI path leaks into the Möbius glyph.
+  assert.doesNotMatch(markup, /M8\.2 7\.2c-2\.9/)
 })
