+++ frontend/src/components/SettingsView/SettingsView.jsx
@@ -41,7 +41,7 @@ import ManageModelsModal from '../ChatView/ManageModelsModal.jsx'
 import UpdateReviewModal from './UpdateReviewModal.jsx'
 import ProviderUsage from './ProviderUsage.jsx'
 import { formatPlanStatus } from './providerUsage.js'
-import { PROVIDER_INFO, PROVIDER_ORDER } from '../ChatView/ChatSettingsPanel.jsx'
+import { PROVIDER_INFO, PROVIDER_ORDER } from '../ChatView/providerRegistry.jsx'
 import '../ui/StatusDot.css'
 import '../ui/ModelSheet.css'
 import './SettingsView.css'
